==> scripts/03_markers_and_annotation/02_annotations_pass1.Rmd <==
---
title: "annotations - rpca"
author: "Kimberly Olney, PhD"
date: "July/31/2025"
output:
  html_document:
    theme: cerulean
    toc: true
    toc_float: true
editor_options: 
  chunk_output_type: inline
---


```{r setup, include=FALSE}
knitr::opts_knit$set(root.dir = "/tgen_labs/jfryer/kolney/LBD_CWOW/LBD_snRNA/scripts/")
setwd("/tgen_labs/jfryer/kolney/LBD_CWOW/LBD_snRNA/scripts/")
source("file_paths_and_colours.R")
color_panel <- dittoColors()
names(color_panel) <- order_cell_type
```

# Read in object
```{r read_object}
# read object
project_ID <- "CWOW_cellbender"
in_rds <- paste0("../rObjects/", project_ID, "_FULL_RNA_joined_normalized.rds")

dataObject <- readRDS(in_rds)

DefaultAssay(dataObject) <- "RNA"
Idents(dataObject) <- "seurat_clusters"
dataObject
```

# Number of nuclie per sample
```{r number_cells}
# Visualize the number of cell counts per sample
data <- as.data.frame(table(dataObject$Sample_ID))
colnames(data) <- c("Sample_ID","frequency")

ncells <- ggplot(data, aes(x = Sample_ID, y = frequency, fill = Sample_ID)) + 
  geom_col() +
  theme_classic() +
  geom_text(aes(label = frequency), 
            position=position_dodge(width=0.9), 
            #vjust=-0.25, 
            hjust = -.025,
            angle = 90) +
  #scale_fill_manual(values = sample_colors) + 
  scale_y_continuous(breaks = seq(0,20000, by = 2000), limits = c(0,20000)) +
  ggtitle("Nuclei per sample") +
  theme(legend.position =  "none") + 
  theme(axis.text.x = element_text(angle = 45, hjust=1))
path <- paste0("../results/nuclei_count/",project_ID, 
               "_cells_per_sample_post_QC_and_doublets_removed")
ncells
saveToPDF(paste0(path, ".pdf"), width = 12, height = 5)

# mean cell count 
mean(data$frequency)

# median cell count 
median(data$frequency)
```

# UMAP by group and sample
```{r dimplot1}
dataObject
DimPlot(dataObject,
        group.by = "group", reduction = "umap.rpca")

DimPlot(dataObject,
        group.by = "Sample_ID", reduction = "umap.rpca")
```

# Unannotated 
### UMAP
```{r unannotated_umap}
# numerical order 
Idents(dataObject) <- factor(dataObject$seurat_clusters, 
                                     levels = as.character(sort(as.numeric(levels(factor(dataObject$seurat_clusters))))))

levels(Idents(dataObject))

ditto_umap <- dittoDimPlot(object = dataObject,
             size = .2,
             reduction.use = "umap.rpca",
             var = "seurat_clusters",
             do.label = TRUE,
             labels.highlight = TRUE)
ditto_umap

path <- paste0("../results/UMAP/",project_ID,
               "_UMAP_unannotated_pass1_clusters")
ditto_umap
saveToPDF(paste0(path, ".pdf"), width = 12, height = 9)
```

### Nuclei count per cluster
```{r count_per_cluster}
count_per_cluster <- FetchData(dataObject,
                               vars = c("ident")) %>%
  dplyr::count(ident) %>%
  tidyr::spread(ident, n)
count_per_cluster

count_melt <- reshape2::melt(count_per_cluster)
colnames(count_melt) <- c("cluster", "number of nuclei")
count_max <- count_melt[which.max(count_melt$`number of nuclei`), ]
count_max_value <- count_max$`number of nuclei`
cellmax <- count_max_value + 15000 # so that the figure doesn't cut off the text
count_bar <- ggplot(count_melt, aes(x = factor(cluster), y = `number of nuclei`, fill = `cluster`)) +
  geom_bar(
    stat = "identity",
    colour = "black",
    width = 1,
    position = position_dodge(width = 0.8)
  ) +
  geom_text(
    aes(label = `number of nuclei`),
    position = position_dodge(width = 0.9),
    vjust = -0.25,
    angle = 45,
    hjust = -.01
  ) +
  theme_classic() + 
  #scale_fill_manual(values = sample_colors) +
  ggtitle("Number of nuclei per cluster") +  xlab("cluster") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  scale_y_continuous(limits = c(0, cellmax))
count_bar
```
## Feature plot
```{r featureplot}
# UMAP showing the expression of select features
path <- paste0("../results/UMAP/",project_ID,
               "_UMAP_feature_celltype_marker_pass1")
pdf(paste0(path, ".pdf"), width = 12, height = 12)
umap_feature <-
  FeaturePlot(dataObject,
              reduction = "umap.rpca",
              keep.scale = NULL, 
              features = c("SNAP25", "GAD1", "PLP1", "OLIG1", "P2RY12", "AQP4", "FLT1", "DCN"))
umap_feature
dev.off()
```
### DotPlot
```{r dot_individual}
dot_ind <- DotPlot(dataObject,
                   features = genes_markers, 
                   cluster.idents = FALSE,
                   dot.scale = 8) + RotatedAxis()


pdf(paste0("../results/DotPlot/",project_ID,
               "_pass1.pdf"), 
    width = 12, height = 7)
dot_ind
dev.off()
```


```{r cleanup}
rm(ncells, umap_feature, ditto_umap, data, count_melt, count_max, count_per_cluster, count_bar)
```

# load markers
```{r load_markers}
markers <- readRDS(paste0("../rObjects/", project_ID,"_markers_presto.rds"))
all.markers.strict <- readRDS(paste0("../rObjects/", project_ID,"_markers_presto_log2FC1_q0.05.rds"))
```

# Get markers for each cluster
```{r markers_per_cluster}
# unique clusters variable
unique_clusters <- unique(markers$cluster)

# empty list to store individual cluster data frames
cluster_list <- list()

# loop through each cluster and create a data frame
for (i in unique_clusters) {
  cluster_name <- paste0("cluster", i)
  cluster_data <- all.markers.strict[all.markers.strict$cluster == i, ]
  assign(cluster_name, cluster_data)
  cluster_list[[cluster_name]] <- cluster_data
}
```

# Cluster Annotation

## Cluster
```{r cluster0}
# Number of cells per condition
count_per_cluster[,c(1,2)]
# UMAP with only cluster 0
DimPlot(object = subset(dataObject, seurat_clusters == "40"),
        reduction = "umap", 
        label = TRUE,
        label.box = TRUE,
        label.size = 3,
        repel = TRUE,
        cols = color_panel[1])


DimPlot(object = subset(dataObject, seurat_clusters == "39"), reduction = "integrated.rpca", group.by = c("group"))


VlnPlot(dataObject,
        features = cluster0$gene[1:10],
        stack = TRUE,
        flip = TRUE,
        split.by = "seurat_clusters")
cluster0$gene[1:10]
```


# Assign identities
## Individual 
```{r assign_individual}
dataObject.annotated <- RenameIdents(object = dataObject, 
                               "0" = "oligodendrocyte",
                               "1" = "oligodendrocyte",
                               "2" = "astrocyte",
                               "3" = "neuron",
                               "4" = "opc",
                               "5" = "microglia",
                               "6" = "interneuron",
                               "7" = "neuron",
                               "8" = "neuron",
                               "9" = "oligodendrocyte",
                               "10" = "interneuron",
                               "11" = "neuron", 
                               "12" = "neuron",
                               "13" = "interneuron",
                               "14" = "interneuron",
                               "15" = "neuron",
                               "16" = "neuron",
                               "17" = "neuron", 
                               "18" = "mural", 
                               "19" = "fibroblast",
                               "20" = "interneuron",
                               "21" = "neuron",
                               "22" = "endothelial"
                               )
dataObject.annotated$cell_type <- factor(Idents(dataObject.annotated))


UMAP_ind <- dittoDimPlot(object = dataObject.annotated,
             size = 0.5,
             var = "cell_type",
             reduction.use = "umap.rpca",
             do.label = TRUE,
             labels.highlight = TRUE)

path <- paste0("../../results/UMAP/",project_ID,
               "_UMAP_annotated_pass1")
pdf(paste0(path, ".pdf"), width = 12, height = 9)
UMAP_ind
dev.off()
```



### Nuclei count per cluster
```{r annotated_counts_individual}
dataObject.annotated$cell_type
count_per_cluster <- FetchData(dataObject.annotated,
                               vars = c("ident")) %>%
  dplyr::count(ident) %>%
  tidyr::spread(ident, n)
count_per_cluster

count_melt <- reshape2::melt(count_per_cluster)
colnames(count_melt) <- c("cluster", "number of nuclei")
count_max <- count_melt[which.max(count_melt$`number of nuclei`), ]
count_max_value <- count_max$`number of nuclei`
cellmax <- count_max_value + 25000 # so that the figure doesn't cut off the text

# Re-order the melted dataframe factor
count_melt$cluster <- factor(count_melt$cluster, levels = order_cell_type)

count_bar <- ggplot(count_melt, aes(x = cluster, y = `number of nuclei`, fill = cluster)) +
  geom_bar(
    stat = "identity",
    colour = "black",
    width = 0.8 # Reduced from 1 for better spacing
  ) +
  geom_text(
    aes(label = `number of nuclei`),
    vjust = -0.5,
    angle = 45,
    hjust = -0.1,
    size = 3
  ) +
  theme_classic() + 
  scale_fill_manual(values = color_panel) + # color.panel must have 9 colors
  scale_y_continuous(limits = c(0, cellmax), expand = expansion(mult = c(0, 0.1))) +
  labs(title = "Number of nuclei per cluster", x = "Cluster") +
  theme(axis.text.x = element_text(angle = 45, hjust = 1))

count_bar
```

```{r nuclei_count_individual, echo=FALSE}
pdf(
  paste0(
    "../../results/UMAP/",
    project_ID,
    "_nuclei_count_annotated_pass1.pdf"
  ),
  width = 6,
  height = 4
)
count_bar
dev.off()
```

### DotPlot
```{r dot_individual}
dot_ind <- DotPlot(dataObject.annotated,
                   features = genes_markers, 
                   cluster.idents = TRUE,
                   dot.scale = 8) + RotatedAxis()
dot_ind
```

```{r save_dot_individual, echo=FALSE}
pdf(
  paste0(
    "../../results/DotPlot/",
    project_ID,
    "_DotPlot_annotation_pass1.pdf"
  ),
  width = 14,
  height = 5
)
dot_ind
dev.off()
```


### Relative abundance
```{r relative_cell_type}
# 1. Calculate global abundance (one row per cell type)
relative_abundance <- dataObject.annotated@meta.data %>%
  group_by(cell_type) %>% # Use your cluster/cell type column
  dplyr::count() %>%
  ungroup() %>%
  dplyr::mutate(percent = 100 * n / sum(n)) %>%
  arrange(desc(percent))

# 2. Plot with a static X-axis label
rel_abun <- ggplot(relative_abundance, aes(x = "Total Dataset", y = percent, fill = cell_type)) +
  geom_col(width = 0.5) + # Adjusted width for a single bar
  geom_text(aes(label = paste0(round(percent, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            size = 3.5, 
            color = "white",
            fontface = "bold") +
  scale_fill_manual(values = color_panel) +
  labs(x = NULL, y = "Percentage (%)", title = "Global Cell Type Composition") +
  theme_bw() +
  theme(panel.grid.major.x = element_blank()) # Clean up the X-axis

rel_abun
pdf(
  paste0(
    "../../results/UMAP/",
    project_ID,
    "_annotated_relative_abundance_pass1.pdf"
  ),
  width = 3,
  height = 5
)
rel_abun
dev.off()
```

```{r}
relative_abundance <- dataObject.annotated@meta.data %>%
  group_by(cell_type, Sample_ID) %>%
  dplyr::count() %>%
  group_by(Sample_ID) %>%
  dplyr::mutate(percent = 100 * n / sum(n)) %>%
  ungroup()


rel_abun <- ggplot(relative_abundance, aes(x = Sample_ID, y = percent, fill = cell_type)) +
  geom_col() +
  geom_text(aes(label = paste0(round(percent), "%")), 
            position = position_stack(vjust = 0.5), size = 3, color = "white") +
  scale_fill_manual(values = color_panel) +
  ggtitle("Percentage of cell type") +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1))
rel_abun

pdf(
  paste0(
    "../../results/UMAP/",
    project_ID,
    "_annotated_relative_abundance_Sample_ID_pass1.pdf"
  ),
  width = 13,
  height = 6
)
rel_abun
dev.off()

# Group
relative_abundance <- dataObject.annotated@meta.data %>%
  group_by(cell_type, group) %>%
  dplyr::count() %>%
  group_by(group) %>%
  dplyr::mutate(percent = 100 * n / sum(n)) %>%
  ungroup()


rel_abun <- ggplot(relative_abundance, aes(x = group, y = percent, fill = cell_type)) +
  geom_col() +
  geom_text(aes(label = paste0(round(percent), "%")), 
            position = position_stack(vjust = 0.5), size = 3, color = "white") +
  scale_fill_manual(values = color_panel) +
  ggtitle("Percentage of cell type") +
  theme_bw() +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust = 1))
rel_abun

pdf(
  paste0(
    "../../results/UMAP/",
    project_ID,
    "_annotated_relative_abundance_group_pass1.pdf"
  ),
  width = 8,
  height = 6
)
rel_abun
dev.off()
```

# Save RDS 
```{r save_annotations}
saveRDS(dataObject.annotated, paste0("../../rObjects/", project_ID, "_annotated_before_subcluster.rds"), compress = FALSE)
```
